Add files via upload
This is the part one of my work about the "Jane Street Real-Time Market Data Forecasting" competition, which has mainly explored the sourcee data offered with the official site. In this file, I have finished the data-truncating and aligning, outliers detected and filled, statistic information combing, features and responders correlation analyzing, and so on. At the end, some important results are recorded to a .xlsx file.